Respond to CR.

+++ SignalServiceKit/src/Messages/Interactions/TSMessage.m
@@ -101,8 +101,8 @@ static const NSUInteger OWSMessageSchemaVersion = 4;
                    attachmentIds:(NSArray<NSString *> *)attachmentIds
                             body:(nullable NSString *)body
                     contactShare:(nullable OWSContact *)contactShare
-                 expireStartedAt:(unsigned long long)expireStartedAt
-                       expiresAt:(unsigned long long)expiresAt
+                 expireStartedAt:(uint64_t)expireStartedAt
+                       expiresAt:(uint64_t)expiresAt
                 expiresInSeconds:(unsigned int)expiresInSeconds
                      linkPreview:(nullable OWSLinkPreview *)linkPreview
                    quotedMessage:(nullable TSQuotedMessage *)quotedMessage
