+++ SignalServiceKit/src/Messages/OWSAddToContactsOfferMessage.h
@@ -25,8 +25,8 @@ __attribute__((deprecated)) @interface OWSAddToContactsOfferMessage : TSInfoMess
                    attachmentIds:(NSArray<NSString *> *)attachmentIds
                             body:(nullable NSString *)body
                     contactShare:(nullable OWSContact *)contactShare
-                 expireStartedAt:(unsigned long long)expireStartedAt
-                       expiresAt:(unsigned long long)expiresAt
+                 expireStartedAt:(uint64_t)expireStartedAt
+                       expiresAt:(uint64_t)expiresAt
                 expiresInSeconds:(unsigned int)expiresInSeconds
                      linkPreview:(nullable OWSLinkPreview *)linkPreview
                    quotedMessage:(nullable TSQuotedMessage *)quotedMessage
